+++ src/model/Calculator.java
@@ -7,10 +7,10 @@ import java.rmi.RemoteException;
  * This interface names the methods that should be implemented by any class that implements this interface.
  * 
  * Named Methods:
- *  int add(int num1, int num2) throws RemoteException;
-	int subtract(int num1, int num2) throws RemoteException;
-	int multiply(int num1, int num2) throws RemoteException;
-	int divide(int num1, int num2) throws RemoteException;
+ *  double add(double num1, double num2) throws RemoteException;
+	double subtract(double num1, double num2) throws RemoteException;
+	double multiply(double num1, double num2) throws RemoteException;
+	double divide(double num1, double num2) throws RemoteException;
 	void calculatorConnected()
 	
 	
@@ -18,9 +18,9 @@ import java.rmi.RemoteException;
  * @version 1.0
  */
 public interface Calculator extends Remote{
-	int add(int num1, int num2) throws RemoteException; //return the sum of two numbers
-	int subtract(int num1, int num2) throws RemoteException; //return the difference of two numbers
-	int multiply(int num1, int num2) throws RemoteException; //return the product of two numbers
-	int divide(int num1, int num2) throws RemoteException; //return the result of a number divided by another
+	double add(double num1, double num2) throws RemoteException; //return the sum of two numbers
+	double subtract(double num1, double num2) throws RemoteException; //return the difference of two numbers
+	double multiply(double num1, double num2) throws RemoteException; //return the product of two numbers
+	double divide(double num1, double num2) throws RemoteException; //return the result of a number divided by another
 	void calculatorConnected() throws RemoteException;  //send a message to the server to notify of a client connected.
 }
